% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crop_seed_info}
\alias{crop_seed_info}
\title{Crop seed info}
\format{A data frame with 18 variables: \code{year}, \code{crop_species},
  \code{seed_vendor}, \code{crop_variety}, \code{glyphosate_tolerance},
  \code{fungicide_azoxystrobin}, \code{fungicide_ethaboxam}, 
  \code{fungicide_fludioxonil}, \code{fungicide_mefenoxam},
  \code{fungicide_metalaxyl}, \code{fungicide_prothioconazole},
  \code{fungicide_penflufen}, \code{fungicide_thiabendazole},
  \code{insecticide_clothianidin}, \code{insecticide_imidacloprid},
  \code{nematicide_bacillusfirmus}, \code{BT_genes}, and \code{notes}}
\usage{
crop_seed_info
}
\description{
Details about crop seed type by year.
}
\keyword{datasets}

