% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRIPSMeta.R
\docType{package}
\name{STRIPSMeta}
\alias{STRIPSMeta}
\alias{STRIPSMeta-package}
\title{STRIPSMeta: A package for STRIPS meta data.}
\description{
The STRIPSMeta package contains STRIPS meta data including watershed
information, e.g. treatment, size, etc., and crop seed details.
}
\section{STRIPSMeta data sets}{

\code{\link{watersheds}}
\code{\link{crop_seed_info}}
}
\seealso{
\code{\link{watersheds}}, \code{\link{crop_seed_info}}
}

